Update patch set 2
Patch Set 2:

(2 comments)

Great work so far! Once the acumos library has been updated and released, I believe a unit test should be added here that exercises the "raw" media type functionality in the model runner. E.g. maybe dump and run a dummy model that counts tokens in text, and we POST a raw string (as opposed to a protobuf message or json object) to the model.

Patch-set: 2
Label: Verified=0