Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> Fails because it attempts to test against Python 3.4.  That's not on the LF build minions anymore.  Please update your tox.ini file.

I think the next error is related to python 3.5 and pytest --> zipp requires py3.6 or higher. Not sure if we need to use older version of pytest to get pytest to work on python 3.5. It isn't clear from there documentation. Maybe another project has noted the version of pytest that works with python 3.5

Patch-set: 3
Label: Code-Review=0