Update patch set 5
Patch Set 5:

looks like this is released, I confirmed it.  the other job failed bcos it tried to push again to test.pypi.org.  That job should not have run, but the job config is deficient.  This is a reason to reorganize this repo as was done in acumos-python-client.  Please consider it.

Patch-set: 5