Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Before we go forward, do we need to maintain compatibility for python 3.5?  The main client library itself was pinned to 3.6 and later only already.  
> 
> (https://github.com/acumos/acumos-python-client/blob/master/acumos-package/setup.py#L68)
> 
> This is fancy footwork to get it working, but wondering if it would just be easier/appropriate to push to the newer version only.

Sure, pip will select the latest version compatible with the running python version. so we can just drop 3.5 compatibility.

We should, however, make a last version compatible with python 3.5 with werkzeug version pinned. then support for 3.5 can be dropped. that way, older version of the python client ( ie, models onboarded with an older version of the client or on 3.5) won't crash.

Patch-set: 2