Update patch set 2
Patch Set 2:

I'd be more comfortable if we just had the minimum required for 3.5, but if we can line up another change to remove this extra code (e.g. assume 3.5 is not supported), I would approve them both.  The global checks and special operations in testing are a bit heavy for long-term project support -- imagine coming back to this in 8 months ;)

Specifically: we can go for 3.5 fix, push/merge, but then we need to have streamlined version beyond 3.5 right behind it.

Patch-set: 2
Reviewer: Gerrit User 10 <10@655300a2-7a56-4504-9153-881edd907bea>