Update patch set 2
Patch Set 2:

Then we should abandon this change and only merge 7785 as it provides a fix for this issue as well (since the CI would not run otherwise) and keeps compatibility with 3.5+ (I'd rather check first if 3.7 would also work). The upgrade to connexion 2 and werkzeug > 1 was just the cherry on top.

Patch-set: 2